% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blsp.R
\name{blsp}
\alias{blsp}
\title{This creates a scatter plot matrix in black and white for the species of
interest}
\usage{
blsp(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
A black and white scatter plot matrix.
}
\description{
This creates a scatter plot matrix in black and white for the species of
interest
}
\examples{
pairs(auris[,2:14],lower.panel=NULL)
Below this is a guide on where to put your information in
the function when writing it out:

}
