% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colspm.R
\name{colspm}
\alias{colspm}
\title{This creates a scatter plot matrix in color for the species of
interest}
\usage{
colspm(x, y)
}
\arguments{
\item{x}{Numeric vector}

\item{y}{Numeric vector}
}
\value{
A scatter plot matrix in color.
}
\description{
This creates a scatter plot matrix in color for the species of
interest
}
\examples{
mycols <- c("red", "green", "blue")
Below this is a guide on where to put your information in
the function when writing it out:

choose any color for the selected variable by typing
out the name of the color:

ex. 'light blue' or 'red' or 'green'

Option 1: simple
}
